Refactor the structure tree
