+++ lib/app/routes/routes.dart
@@ -1,6 +1,5 @@
 import 'package:expense_tracker/app/app.dart';
-import 'package:expense_tracker/presentations/pages/authentication/forgot_password/forgot_pw_page.dart';
-import 'package:expense_tracker/presentations/pages/home/components/transaction_card.dart';
+import 'package:expense_tracker/presentations/pages/forgot_password/forgot_pw_page.dart';
 import 'package:expense_tracker/presentations/pages/home/home_page.dart';
 import 'package:expense_tracker/presentations/pages/login/login_page.dart';
 import 'package:expense_tracker/presentations/pages/profile/account/account_page.dart';
@@ -48,5 +47,5 @@ final Map<String, WidgetBuilder> routes = {
   SettingPage.routeName: (context) => const SettingPage(),
   ExportPage.routeName: (context) => const ExportPage(),
   ExportDataPromptPage.routeName: (context) => const ExportDataPromptPage(),
-  TransactionDetailPage.routeName: (context) => const TransactionDetailPage(),
+  // TransactionDetailPage.routeName: (context) => const TransactionDetailPage(),
 };
