--- lib/infrastructure/core/firebase_injectable_module.dart
@@ -1,15 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:google_sign_in/google_sign_in.dart';
-import 'package:injectable/injectable.dart';
-
-/// Wrapper for instantiate 3rd library in get_it
-@module
-abstract class FirebaseInjectableModule {
-  @lazySingleton
-  GoogleSignIn get googleSignIn => GoogleSignIn();
-  @lazySingleton
-  FirebaseAuth get firebaseAuth => FirebaseAuth.instance;
-  @lazySingleton
-  FirebaseFirestore get firestore => FirebaseFirestore.instance;
-}
