+++ lib/injector.dart
@@ -1,6 +1,9 @@
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:google_sign_in/google_sign_in.dart';
 
 import 'injector.config.dart';
 
@@ -14,4 +17,15 @@ void configureInjection(String env) {
 abstract class SharedPreferencesModule {
   @preResolve
   Future<SharedPreferences> get prefs => SharedPreferences.getInstance();
+}
+
+/// Wrapper for instantiate 3rd library in get_it
+@module
+abstract class FirebaseInjectableModule {
+  @lazySingleton
+  GoogleSignIn get googleSignIn => GoogleSignIn();
+  @lazySingleton
+  FirebaseAuth get firebaseAuth => FirebaseAuth.instance;
+  @lazySingleton
+  FirebaseFirestore get firestore => FirebaseFirestore.instance;
 }
