+++ lib/presentations/pages/home/widgets/widgets.dart
@@ -1,1 +1,3 @@
 export 'avatar.dart';
+export 'line_chard_widget.dart';
+export 'top_navigation.dart';
