+++ lib/presentations/pages/login/login_form.dart
@@ -1,6 +1,6 @@
 import 'package:expense_tracker/blocs/login/login_cubit.dart';
 import 'package:expense_tracker/presentations/components/default_button.dart';
-import 'package:expense_tracker/presentations/pages/authentication/forgot_password/forgot_pw_page.dart';
+import 'package:expense_tracker/presentations/pages/forgot_password/forgot_pw_page.dart';
 import 'package:expense_tracker/presentations/pages/login/widgets.dart';
 import 'package:expense_tracker/signup/views/signup_page.dart';
 import 'package:flutter/material.dart';
